TEMPORARY: Incomplete eslint stuff from previous PC
Changes recovered from the PC I used to use for working on this project.
These are incomplete and should not be merged without review. There are
probably further changes to be made, multiple commits to split this
into, and possible conflicts with current master (these changes are
almost a year old at time of commit).